+++ test-support/helpers/ember-power-select.js
@@ -7,7 +7,7 @@ function typeText(selector, text) {
   $(selector).trigger('input');
 }
 
-function nativeMouseDown(selector, options = {}) {
+export function nativeMouseDown(selector, options = {}) {
   let event = new window.Event('mousedown', { bubbles: true, cancelable: true, view: window });
   Object.keys(options).forEach(key => event[key] = options[key]);
   Ember.run(() => Ember.$(selector)[0].dispatchEvent(event));